Move QubesArgumentParser from tools to toolparsers
